+++ src/Events/StatusUpdated.php
@@ -2,38 +2,26 @@
 
 namespace Spatie\ModelStatus\Events;
 
+use Spatie\ModelStatus\Status;
 use Illuminate\Database\Eloquent\Model;
 
-/**
- * Event fired by the HasStatus trait when a status is updated
- *
- * @package Spatie\ModelStatus\Events
- */
 class StatusUpdated
 {
-    protected $oldStatus;
-    protected $newStatus;
-    protected $model;
+    /** @var \Spatie\ModelStatus\Status|null */
+    public $oldStatus;
 
-    public function __construct(string $oldStatus, string $newStatus, Model $model)
+    /** @var \Spatie\ModelStatus\Status */
+    public $newStatus;
+
+    /** @var \Illuminate\Database\Eloquent\Model */
+    public $model;
+
+    public function __construct(?Status $oldStatus, Status $newStatus, Model $model)
     {
         $this->oldStatus = $oldStatus;
+
         $this->newStatus = $newStatus;
+
         $this->model = $model;
-    }
-
-    public function getOldStatus(): string
-    {
-        return $this->oldStatus;
-    }
-
-    public function getNewStatus(): string
-    {
-        return $this->newStatus;
-    }
-
-    public function getModel(): Model
-    {
-        return $this->model;
     }
 }