+++ tests/HasStatusesTest.php
@@ -2,11 +2,10 @@
 
 namespace Spatie\ModelStatus\Tests;
 
-use Spatie\ModelStatus\Exceptions\InvalidStatus;
-use Spatie\ModelStatus\Tests\Models\AlternativeStatusModel;
-use Spatie\ModelStatus\Tests\Models\CustomModelKeyStatusModel;
 use Spatie\ModelStatus\Tests\Models\TestModel;
+use Spatie\ModelStatus\Exceptions\InvalidStatus;
 use Spatie\ModelStatus\Tests\Models\ValidationTestModel;
+use Spatie\ModelStatus\Tests\Models\AlternativeStatusModel;
 
 class HasStatusesTest extends TestCase
 {
@@ -215,6 +214,16 @@ class HasStatusesTest extends TestCase
         $this->assertCount(4, TestModel::otherCurrentStatus('initiated')->get());
         $this->assertCount(3, TestModel::otherCurrentStatus('initiated', 'pending')->get());
         $this->assertCount(3, TestModel::otherCurrentStatus(['initiated', 'pending'])->get());
+    }
+
+    /** @test */
+    public function it_supports_custom_polymorphic_model_types()
+    {
+        Relation::morphMap(['custom-test-model' => TestModel::class]);
+
+        $this->testModel->setStatus('initiated');
+
+        $this->assertCount(1, TestModel::currentStatus('initiated')->get());
     }
 
     /** @test */